move model-subscription projector
